+++ flask_unchained/bundles/sqlalchemy/hooks/__init__.py
@@ -1,2 +1,1 @@
-from .model_fixture_folders_hook import ModelFixtureFoldersHook
 from .register_models_hook import RegisterModelsHook